proto-ext-whiterabbit/wr-servo: fix the overflow of casted picos_mu
For long liks (like 180km) "(signed)s->picos_mu" overflows causing wrong
result of the comparison:
(signed)s->picos_mu < (signed)big_delta_fix)

+++ proto-ext-whiterabbit/wr-servo.c
@@ -276,7 +276,7 @@ int wr_servo_update(struct pp_instance *ppi)
 	big_delta_fix =  s->delta_tx_m + s->delta_tx_s
 		       + s->delta_rx_m + s->delta_rx_s;
 
-	if ((signed)s->picos_mu < (signed)big_delta_fix) {
+	if ((int64_t)s->picos_mu < (int64_t)big_delta_fix) {
 		/* avoid negatives in calculations */
 		s->picos_mu = big_delta_fix;
 	}
